thesis_formatter_complete/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.0.1
lxml==4.9.3
openpyxl==3.1.2
requests==2.31.0